clearing styles
