+++ src/icons/index.js
@@ -19,6 +19,7 @@ export { default as folderProgram16 } from "./folderProgram16-min.png";
 export { default as folderProgram24 } from "./folderProgram24-min.png";
 export { default as help24 } from "./help24-min.png";
 export { default as htmlFile16 } from "./htmlFile16-min.png";
+export { default as htmlFile32 } from "./htmlFile32-min.png";
 export { default as imaging16 } from "./imaging16-min.png";
 export { default as internetExplorere16 } from "./internetExplorer16-min.png";
 export { default as internetExplorere32 } from "./internetExplorer32-min.png";
